+++ tests/TypingErrorTest.php
@@ -1,0 +1,55 @@
+<?php
+/**
+ * @package axy\errors
+ */
+
+namespace axy\errors\tests;
+
+use axy\errors\TypingError;
+
+/**
+ * @coversDefaultClass axy\errors\TypingError
+ */
+class TypingErrorTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @covers ::__construct
+     * @covers ::getVarname
+     * @covers ::getExpected
+     */
+    public function testNormal()
+    {
+        $previous = new \RuntimeException('msg');
+        $e = new TypingError('var', 'array', $previous);
+        $this->assertSame('var', $e->getVarname());
+        $this->assertSame(['array'], $e->getExpected());
+        $this->assertSame($previous, $e->getPrevious());
+        $this->assertSame('var must be array', $e->getMessage());
+    }
+
+    /**
+     * @covers ::__construct
+     * @covers ::getVarname
+     * @covers ::getExpected
+     */
+    public function testExpectedArray()
+    {
+        $e = new TypingError('arg', ['MyClass', 'array', 'string']);
+        $this->assertSame('arg', $e->getVarname());
+        $this->assertSame(['MyClass', 'array', 'string'], $e->getExpected());
+        $this->assertSame('arg must be MyClass, array or string', $e->getMessage());
+    }
+
+    /**
+     * @covers ::__construct
+     * @covers ::getVarname
+     * @covers ::getExpected
+     */
+    public function testDefault()
+    {
+        $e = new TypingError();
+        $this->assertNull($e->getVarname());
+        $this->assertSame([], $e->getExpected());
+        $this->assertSame('A value must be a different type', $e->getMessage());
+    }
+}